# openapi.yaml
openapi: 3.0.3
info:
  title: Rock LMS Course Management API
  version: 1.0.0
  description: API for managing Rock LMS courses (CRUD + publish/archive)

servers:
  - url: http://localhost:8080/api

paths:
  /courses:
    get:
      operationId: findAllCourses
      summary: Retrieve courses with optional status filtering
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [DRAFT, PUBLISHED, ARCHIVED]
          required: false
          description: Filter courses by status
      responses:
        '200':
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CourseDto'
    post:
      operationId: createCourse
      summary: Create new course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreateRequest'
      responses:
        '201':
          description: Course created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'

  /courses/{id}:
    get:
      operationId: findCourseById
      summary: Retrieve course by identifier
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
        '404':
          description: Course not found
    put:
      operationId: updateCourse
      summary: Modify existing course
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseUpdateRequest'
      responses:
        '200':
          description: Course updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
        '400':
          description: Cannot modify archived course
    delete:
      operationId: removeCourse
      summary: Remove course permanently
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Course removed successfully

  /courses/{id}/publish:
    post:
      operationId: publishCourse
      summary: Make course available to students
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'
        '400':
          description: Course missing required fields

  /courses/{id}/archive:
    post:
      operationId: archiveCourse
      summary: Move course to archived state
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Course archived successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CourseDto'

components:
  schemas:
    CourseDto:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
          minimum: 1
        status:
          type: string
          enum: [DRAFT, PUBLISHED, ARCHIVED]
        publishedAt:
          type: string
          format: date-time
          nullable: true

    CourseCreateRequest:
      type: object
      required: [title, duration]
      properties:
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
          minimum: 1

    CourseUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        duration:
          type: integer
          minimum: 1